SDK-909: Self review code fixes.
